Force type casting to return 0 if winner check is false.

+++ application/controllers/board.php
@@ -170,7 +170,7 @@ class Board extends CI_Controller {
 
   /* Check if a player has won */
   function check_if_winner() {
-    echo json_encode (array('status'=> check_horizontal() || check_vertical() || check_diagonal()));
+    echo json_encode (array('status'=> (int)(check_horizontal() || check_vertical() || check_diagonal() )));
     return;
   }
